Add support for logging "partial requests" which never finish
This introduces configurable "partial" logfiles, which are per-process
logfiles which should only ever have 0 or 1 request in them.  As soon
as possible in the request, we write out the information we have to
these files -- as soon as the request is successful and written to the
permanent logfile, we rewind and truncate the log message.  This means
that any non-empty files pinpoint requests which were never completed,
and bear examination.